Fix issue with a warning that it triggers now but now before, whereas the code has not changed. Another random warning

+++ m-core.h
@@ -2099,7 +2099,7 @@ m_core_hash (const void *str, size_t length)
 #define M_HASH_INT32(a) ( (a) ^ ((a) << 11) ^ M_HASH_SEED )
 #define M_HASH_INT64(a) ( ( (a) >> 33 ) ^ (a) ^ ((a) << 11) ^ M_HASH_SEED )
 #define M_HASH_DEFAULT(a)                                               \
-  _Generic((a)+0,                                                       \
+  (size_t) _Generic((a)+0,                                              \
            int32_t:  M_HASH_INT32((uint32_t) M_AS_TYPE(int32_t, a)),    \
            uint32_t: M_HASH_INT32(M_AS_TYPE(uint32_t, a)),              \
            int64_t:  M_HASH_INT64((uint64_t) M_AS_TYPE(int64_t, a)),    \
